Fix unchecked use of BuildContext after await
When using the BuildContext after an asynchronus gap (like when using
await), you always need to check if the context is still mounted before
using it to retrieve objects from the widget tree like with
`ScaffoldMessenger.of(context)`.

Additionally you don't need to use
`.then((value) {})` on a future, if that future is already awaited, you
can simply put any following statements on a new line for better
readability of program flow.