Move copy message snackbar to top of view
On Android 13 or newer there is an operating system integrated message
bubble that pops up and shows you any automated clipboard changes at the
bottom left of the screen. This bubble obscures the message text of the
snackbar so I've moved it to the top of the screen when on Android.

It might make sense to do more accurate version checking in the future
to only show the snackbar if on Android < 13 or iOS but we'd need to pull in the
device_info_plus package or similar for that.